Handle the case where the is no resolver
Some python packages, e.g. pytz, are purely written in python and do not
have any other dependencies. In those cases, a `resolver` is not set on
the compiler, causing `AttributeError` exceptions in various places.

This commit performs a `None` check on the `resolver` attribute before
proceeding to access nested attributes.

Fixes: #304

Signed-off-by: Luiz Carvalho <lucarval@redhat.com>